fix: correct data types for ibl_plr columns to match native type handling
## Schema Changes

- Change `retired` from VARCHAR to TINYINT(1)
- Change `cy` and `cyt` from INT to TINYINT UNSIGNED
- Change `cy1-cy6` from INT to SMALLINT
- Change `coach` from VARCHAR to TINYINT UNSIGNED
- Reduce `teamname` from VARCHAR(32) to VARCHAR(16)
- Reduce `draftedby` and `draftedbycurrentname` from VARCHAR(32) to VARCHAR(16)
- Reduce `college` from VARCHAR(48) to VARCHAR(40)

## PHP Code Updates

- Update all `retired` comparisons from `=== '0'` to `=== 0` (int)
- Update Player/PlayerData `isRetired` property type from `?string` to `?int`
- Update PlayerStats `isRetired` property type from `string` to `int`
- Remove unnecessary `(int)` casts on native INT columns (pid, tid, etc.)
- Update PHPDoc type hints for `retired` field (string → int)
- Update CashTransactionHandler `cyt` type from string to int

## Migration

- Add migration 011_ibl_plr_type_corrections.sql to apply schema changes

## Rationale

MYSQLI_OPT_INT_AND_FLOAT_NATIVE is enabled, so INT/TINYINT columns return PHP int types. Schema now uses correct column types: TINYINT for small numbers (0-6 year contracts, boolean flags), SMALLINT for salary values (0-9999), and appropriately-sized VARCHAR for text fields based on actual data constraints.

+++ ibl5/classes/ContractList/ContractListService.php
@@ -54,7 +54,7 @@ class ContractListService implements ContractListServiceInterface
                 'name' => $player['name'],
                 'pos' => $player['pos'],
                 'teamname' => $player['teamname'],
-                'tid' => (int) ($player['tid'] ?? 0),
+                'tid' => $player['tid'] ?? 0,
                 'team_city' => $player['team_city'] ?? '',
                 'color1' => $player['color1'] ?? 'FFFFFF',
                 'color2' => $player['color2'] ?? '000000',
